pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darwin-perf"
version = "0.2.1"
description = "System performance monitoring for macOS Apple Silicon — GPU, CPU, memory, energy, disk I/O via Mach APIs. No sudo needed."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Adam Mikulis"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: C",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["macos", "darwin", "gpu", "cpu", "monitoring", "performance", "apple-silicon", "metal", "mlx", "pytorch", "mps", "m1", "m2", "m3", "m4", "m5"]

[project.optional-dependencies]
tui = ["textual>=0.40"]
gui = ["pywebview>=5.0"]
all = ["textual>=0.40", "pywebview>=5.0"]

[project.scripts]
darwin-perf = "darwin_perf.cli:main"

[project.urls]
Homepage = "https://github.com/adammikulis/darwin-perf"
Issues = "https://github.com/adammikulis/darwin-perf/issues"

[tool.setuptools.packages.find]
where = ["src"]
